Add support for explicitly naming WASM runtime library
In some cases the runtime is named differently from the library (for
eg WAMR & libiwasm.so). The current build system assumes the runtime &
library are named the same.

Introduced $ngx_wasm_runtime_lib_name to allow explicit naming of
runtime library. If $ngx_wasm_runtime_lib_name is not set (for eg in config),
we fallback to the old behaviour.

    fixes #57