feat(wasi) begin a minimal WASI implementation
This is needed by VMs that do not provide WASI functions, such as V8. Only the
minimal set of functionality required in order to load the .wasm files used by
the test suite is implemented, but some pointers on how to continue the
implementation of environment functions are also provided.