fix(wavm) treat address 0 as NULL
The proxy-wasm-rust-sdk implementation oftranslates the None argument of
set_property to a null pointer:

https://github.com/proxy-wasm/proxy-wasm-rust-sdk/blob/fb7c953adcba19adb79b4e1885cbb532a2fbad22/src/hostcalls.rs

This arrives to our host calls as a ngx_wavm_ptr_t of value zero.

I could not find any authoritative source stating that a zero pointer is an
invalid ("null") pointer in Wasm, and in fact I found comments stating
otherwise [1] (which makes sense, since in bare metal this is not a property
of an ISA, but of the language API/ABIs built on top -- C guarantees NULL to
be `(void *) 0` even if that isn't the case in the architecture). However,
I've also found a comment [2] from the WAMR developer stating that offset 0 is
reserved in WASI-based toolchains, so 0-pointers should be safe to use as NULL
for proxy-wasm.

See:
[1] https://stackoverflow.com/questions/62042733/can-a-rust-pointer-to-heap-allocated-memory-be-0-when-compiled-to-wasm
[2] https://github.com/bytecodealliance/wasm-micro-runtime/issues/537