chore(test) new TEST_NGINX_CLEAN_LOG environment variable
When set, empties the contents of nginx logs/ *before* running the
tests, which can come in handy when inspecting error logs while running
a test in isolation over and over again (else it is difficult to discern
when the previous tests ends and when the new one beings). It's also
better than removing and re-introducing the file since some tools don't
support runtime file system changes.