chore(release) install CMake >= 3.12 on Ubuntu 18.04
The latest cargo uses cmake --parallel when building
wabt-sys (which is required by cwabt, used by the V8 backend).
The --parallel flag is only available from CMake 3.12 onwards,
which is not included in the Ubuntu 18.04 distro.

More info on the cmake version requirement for cmake-rs:
https://github.com/rust-lang/cmake-rs/issues/131
A fix was merged but a new version of the crate was not
yet released:
https://github.com/rust-lang/cmake-rs/pull/146

Additionally, wabt-sys actually includes code which assumes
CMake >= 3.12! See:
https://ma.ttias.be/wabt-sys-compile-error-cmake-project-version-not-declared-in-scope-ubuntu-18-04/