fix(proxy-wasm) store real content length to trigger body callbacks
We need the original content length for checking whether to trigger the
execution of the on_request_body callback and to trigger the the eof status of
on_response_body callback, in the event that a previous callback
(on_request_headers or on_response_headers) has edited the Content-Length
header.

This change was necessary in order to trigger the callback behavior correctly
in the http_body example from proxy-wasm-go-sdk.