fix(wrt) report error on missing WASI functions (V8 and Wasmer)
Prior to this, on the event of unimplemented WASI functions,
we'd only get a crash caused by the failed assertion at
`ngx_wasm_assert(module->nimports == module->import_types->size)`
and no error message.

This is especially useful on V8 since unimplemented functions
are NYIs from our end; it's less likely on Wasmer but future
unknown WASI versions could trigger this as well.

It's interesting to note that with this change, the behavior of
the Wasmer and V8 backends now matches that of Wasmtime (note that
Test 15 in `t/01-wasm/002-module_directive.t` is reenabled with
no changes to the Wasmtime backend). However, Wasmtime doesn't
report on individual missing WASI functions at link time, and
thus the new Test 16 is disabled for Wasmtime only.

Co-authored-by: Thibault Charbonnier <thibaultcha@users.noreply.github.com>