fix(proxy-wasm) only run on_response_trailers on resp eof
This is a temporary fix that came up on the FFI branch when I noticed
this handler was called several times for responses that had a body, and
I thought that was strange. Here is a test-case and a fix for main and
the FFI branch.

The bigger conversation revolves around our implementation of
on_response_trailers: as per the SDK spec, it is to be invoked on
*upstream* response trailers, but we our tests only cover
locally-produced responses. We should look at Envoy's behavior in this
case because if we have to run this handler _before_ trailers are
produced by `ngx_http_next_body_filter`, we might need to revisit our
implementation strategy. Tests with ngx_http_proxy_module are also
needed to guarantee behavior in the Gateway.