chore(util) fix 'make clean' target
In 05a76b228b4a4a8bb185b9703c0728f4bcdf12bc we changed the build system
to make the Nginx source an intermediate source tree before the patched
Nginx source, which is then built.

This changes the `make clean` target to do so in the patched source
tree, and not the original one.