fix(proxy-wasm) handle get_property() on unset upstream values
When using another content handler such as `echo`, the `r->upstream`
attribute of a request is not set, which used to cause a segfault in
these properties getters. Since the test was disabled on non-IPv6
machines and GitHub Actions this was uncaught until I just switched to
an IPv6 machine for the holidays.