fix(proxy-wasm) do not log error on "property not found"
Avoid logging error on simple "property not found" cases, where our properties
API returns NGX_DECLINED, which gets then translated to NotFound in the
proxy-wasm API.

Instead, let the caller code, such as filter code or FFI client, deal with
that case and report or handle it.

In the test cases, we replace those error log checks with info log checks
issued by the client code (our hostcalls test filter).

(Note that the Rust SDK does not allow the user to handle a "not found" in
`set_property` operation and merely panics. We do the right thing on our side
and return the NotFound code which is consumed by other SDKs, but since we're
using the Rust SDK on our tests, we check for the panic including the numeric
representation of the unhandled status code.)