feat(*) add 'flags' directives for configuring wasm runtimes
This commit introduces the directive `flag name=value` and runtime
blocks `wasmtime{}`, `wasmer{}` and `v8{}`. The directive `flag` can be
declared in a runtime block to change it's configuration:

    wasm {
        wasmtime {
            flag debug_info on;
            flag static_memory_maximum_size 10m;
        }

        wasmer {
            flag wasm_bulk_memory on;
        }

        v8 {
            flag turbo_stats_wasm on;
        }
    }

A list of supported flags along with handlers were added to each
runtime. If a user tries to set an unsupported or unknown flag the nginx
configuration fails and the flag is reported as the cause.

It's worth noting that many V8 wasm flags completely breaks nginx
leaving a message "stack frames were collapsed".

Similarly, the flag `profiler` can't currently be set to `vtune`.  This
is because flags are applied before the wasm engine is created and if
`profiler` is set to `vtune` before the engine is created, nginx
crashes.

Signed-off-by: Thibault Charbonnier <thibaultcha@me.com>