refactor(*) move filter chain isolation to its proper place
Move filter isolation out of initialization code, which was confusing
and not finalized. Filter isolation is given as an argument to the
filter chain when it is started, which feels more appropriate. Use an
unused C structure which was intended to receive such "filter chain
arguments/context".

Thanks to the above change, filter chain callers can fallback to the
configured `proxy_wasm_isolation` setting, or override it. This commit
also extends `proxy_wasm.attach` with the option to do so.