chore(*) add valgrind_track_register_updates directive to Test::Nginx
If a test triggers a SIGFPE, that causes Valgrind to crash the process,
because it does not track the state of the FP registers carefully by
default.

Or at least it didn't back in 2004, but given the behavior we are
observing, it seems to still be the case. See
 https://sourceforge.net/p/valgrind/mailman/message/10333044/

> > Should I create a bug report for this problem?
>
> You are welcome to do so, but I don't think it's fixable. It's even
> harder than filling in the right FP state (which we don't do) or allowing
> the state to be modified (which we don't do). There are already bugs
> for those things...

This is what the Valgrind documentation has to say about signals:

> If you're using signals in clever ways (for example, catching SIGSEGV,
> modifying page state and restarting the instruction), you're probably
> relying on precise exceptions. In this case, you will need to use
> `--vex-iropt-register-updates=allregs-at-mem-access` or
> `--vex-iropt-register-updates=allregs-at-each-insn`.

(From https://valgrind.org/docs/manual/manual-core.html#manual-core.signals)

Since this is a non-default behavior for Valgrind, which is
possibly costly, we will only enable it on tests that need it.

tests(*) mark test as requiring Valgrind to track updates

This test triggers an intentional SIGFPE under Wasmtime, which
causes Valgrind to crash the process if Valgrind does not track
the state of the FP registers carefully. Since this is a
non-default behavior for Valgrind, we will only enable it on tests
that need it.