refactor(shm) store key, value and node in a single allocation
Keys and values in the shm-based key-value store are byte-aligned
strings, so store them in a single slab allocation, concatenating
the node, key and value data. The node stores pointers to
offsets within the slab-allocated buffer, so the remaining
interfaces remain the same (e.g. nodes work as rbtree nodes
vith regular ngx_str_t key lookups).

The main motivation for this refactor is to make each KV
operation require a single slab allocation, to simplify the
handling of allocation failures as we add more elaborate
eviction logic.

A nice side-effect is potentially more efficient allocation,
due to lower fragmentation.

When values are updated, the node only gets reallocated if the
value is longer in bytes than the previous value.