feat(shm) new 'SLRU' eviction policy
Introduces a new eviction policy for shm-based key-value stores: SLRU
(Slot-aware LRU? Slab-aware LRU? Smart LRU?)

It minimizes deallocations, by keeping individual queues for objects of
the same (binary) order of magnitude in size, which is the same strategy
used by the Nginx slab allocator when allocating slabs into slots.
Within each queue, a LRU policy is used.

In practice, this means that when a KV store is used to store multiple
types of objects of various sizes, a full memory zone will be most
likely to deallocate the least-recently-used objects of the same type,
freeing up an entry in a full slab for that type's size, rather than
unsuccessfully eallocating many small objects until an entire slab of an
incompatible size is freed and reassigned to the necessary size.

When the LRU queue for the item's size is empty, the SLRU algorithm will
attempt the next larger size slot, in order to minimize the number of
elements deallocated. If no larger slot is available, the queues for the
largest smaller sizes will be attempted in order.