chore(release) rename --bin-all to --all
A matter of taste, but that flag builds both source and binary packages, and
is "more powerful" than just "for i in * do --bin", so `--all` feels more
appropriate. `--bin-all` still works for backwards compat / muscle memory.

Plus a minor internal refactor to make it easier for the next person to spot
that those flags are mutually exclusive.