fix(proxy_wasm) accept unquoted NGX_WASM_HOST_PROPERTY_NAMESPACE var
This is not really a bug in our end, but Nginx's build system
doesn't like `--with-cc-opt` entries that feature double-quoted variables
as in `"--with-cc-opt=\"-DNGX_WASM_HOST_PROPERTY_NAMESPACE=\\\"kong\\\"\""`.

This _does_ work when expanding the flag and building our module,
but the compilation of Nginx itself later fails with:

```
src/core/nginx.c:456:49: note: in expansion of macro 'NGX_CONFIGURE'
  456 |         ngx_write_stderr("configure arguments:" NGX_CONFIGURE NGX_LINEFEED);
      |                                                 ^~~~~~~~~~~~~
src/core/nginx.c:456:25: note: to match this '('
  456 |         ngx_write_stderr("configure arguments:" NGX_CONFIGURE NGX_LINEFEED);
      |                         ^
```

So let's just play nice and drop the quotes in our end.