fix(wrt/wasmtime) do not override 'WASMTIME_BACKTRACE_DETAILS' if specified
With `backtraces on`, `WASMTIME_BACKTRACE_DETAILS` is forcefully
enabled.

With `backtraces off` (default), `WASMTIME_BACKTRACE_DETAILS` still
defaults to `0` but can be overridden if specified in the `env`
directive.

This commit also ensures the `env` directives in `TestWasm.pm` are
actually injected, although it disables the injection since no
environment variable is presently injected.