fix(proxy-wasm) use pwctx's own pool on every isolation mode
We were observing linear memory growth when running requests in low-isolation
modes.

The root cause comes from the fact that small allocations had instance-long
lifetimes (including, for example, the arrays for `args` and `rets` for each
Wasm call, causing the observable linear growth when running wrk).

Turns out that `ngx_pfree`, which was being used in the code to free `args` and
`rets`, does not free small allocations. It only frees large allocations; small
allocations are only freed when the pool itself is destroyed.

The solution is to use a request-lived pool for `pwctx`, so we don't hoard
small request-specific allocations over the lifetime of an instance.

We also change the setup of the `pwexec` pool so that it uses the `pwctx`
pool under all isolation modes, like it does for logs. Wasm isolation mode
should pertain only to the Wasm-accessible store, not to the temporary C-side
data structures, which we now free on each request.