fix(proxy-wasm) resume the filter chain on dispatch response
The dispatch could be during `on_request_body`, and when the response is
received, subsequent filters' `on_request_body` steps should also be
invoked.

This is a rudimentary fix for the time being, but now that the
proxy-wasm runloop is fully re-entrant, the resume model could also be
refactored into something cleaner and reusable for other C resume
handlers.

Co-Authored-By: Caio Ramos Casimiro <caiorcasimiro@gmail.com>