feat(proxy-wasm) host-managed property getters and setters
This feature adds support for host-managed property getter and
setter functions.

These feature is implemented in three stages:

* `ngx_proxy_wasm_properties.c`: in our implementation for
  proxy-wasm properties, we now allow the host to specify a
  function which gets called whenever proxy-wasm receives a
  `set_property` or `get_property` call matching the host-defined
  namespace. Those getter and setter functions are then responsible
  to return and store the values, and are able to produce any
  desired side-effects in the host.

* `ngx_wasm_lua_ffi.c`: We then expose this functionality in our
  Lua FFI interface. This is a lightweight wrapper, which accepts
  a C function pointer directly, and associates this function
  to the request context object.

* `lib/resty/wasmx/proxy_wasm.lua`: Finally, we expose a Lua API
  for setting a Lua functions as setters and getters. The interface
  is such that the host Lua program should set one setter function
  and/or getter function, and then within that function properly
  dispatch the behavior based on the received key. Internally, these
  functions use the fact that LuaJIT is able to convert a Lua
  function into a C function pointer. Due to lifetime management,
  those generated C callbacks are a limited resource in the LuaJIT
  VM, so we produce a single pair of these and then have them
  call the user-defined Lua function, which remains as a normal
  Lua closure. (See http://luajit.org/ext_ffi_semantics.html for
  more details on LuaJIT FFI C callback managment.)

To ensure that the lifetime of strings produced by the host continue
valid by the time the getter callback finished running and the Wasm
engine reads the properties, getter/setter-produced values continue
to be stored in the rbtree for host properties that we already had
for plain `get_property`/`set_property`, introduced in da820875ed7.

Since we need to store this data on the C side anyway, we also
add an extra feature as an optimization: the concept
of "const" host-properties. If the Lua callback returns a truthy
third-argument, that signals that that value should be flagged as
constant. Further calls to `get_property` during the same request
will get the property value directly from the ngx_wasm_module rbtree
"cache" without hitting the getter callback again. Test cases in the
test suite demonstrate this feature.

In this PR, the host namespace continues to be statically defined
at compile time via the NGX_WASM_HOST_PROPERTY_NAMESPACE define.
The regular FFI functions for `set_property` and `get_property`
introduced in 447ef130a9 also continue to work normally.