tests(*) switch Valgrind tests from 'opt-out' to 'opt-in'
Remove `--- skip_valgrind` block and helper in favor of an opt-in
mechanism for Valgrind test cases. This allows us much finer control
since we want to ignore most tests (many are redundant for Valgrind) but
a select few, so our CI pipeline can run as quickly as possible and
focused on essential tests.

Also update `TEST_NGINX_RANDOMIZE` to be disabled after the first run,
to inspect or debug re-runs more easily while keeping the tests as fast
as possible when everything is good.

Since the new plan mechanism for TestWasm suites stores the number of
tests, also pass it to existing helper extensions like `---
skip_no_debug`.