tests(*): match_error_log block and match_error_log_grep modifier
A new block handler, `match_error_log` where you give it an array of plain
substrings (not regexes), and the test passes if and only if all occurrences
of the given substrings in the error log match the given list in the exact
order, including repeats.

The subset of relevant lines to be checked can be filtered by a regex using
`match_error_log_grep`. If not given, the entire error.log is checked.