fix(proxy-wasm) always pass 'eof=false' in header steps
Reflect the fact that we do not know whether or not a request/response
has a body in the RequestHeaders and ResponseHeaders steps due to
internal Nginx behavior.

Thus, always pass `eof=false` to these handlers so filters can assume
the same behavior: assuming that the request/response *may or may not*
have a body. As such, both request/response body steps *may or may not*
be invoked (same as Envoy), and the `eof` flag is not an appropriate
heuristic for determining body steps invocations.

This also factorizes the on_http_phases suite into a much more concise
set of tests.

See #504