chore(ci-large) increase `tick_period` in intermittently failing tests
The previous short value was causing the failure of a few tests
that relied on a dispatched call during `on_tick`.

Apparently, the frequent `on_tick` calls overwhelmed Nginx to the
point of triggering an early termination of those tests, due to timeout.

This behaviour was only observed when using V8.