fix(proxy-wasm) consistent context checks for header setters
This commit creates a helper function `ngx_proxy_wasm_hfuncs_set_header_check`
which is used by all proxy-wasm host functions which set headers (set, add,
replace, remove), ensuring a consistent behavior across all of them.

In particular, some of the header setters did not produce `BAD_ARGUMENT` when
`map_type` is invalid.

This also changes the behavior of the functions to produce traps when they
are called in invalid contexts, instead of logging an error and returning `OK`.
This matches the behavior of the body setter functions.