feat(proxy-wasm) metrics
This commit adds support for storing metrics in WasmX shared-memory
key-value store facility.

The workflow users are expected to perform follows from Proxy-Wasm metrics ABI
itself: users define metrics before using them; when a metric is defined a
numeric ID is returned which can later be used for reading or updating its
respective metric. If the system is out of metrics memory when defining a new
metric, the metric definition fails as eviction support hasn't been implemented.

The implemented design, described at [1], allows users to perform most metric
updates without synchronizing Nginx workers, i.e. without the aid of locks.

Users can refer to [2] for a description of how metrics are represented in
memory and how to estimate the size of the shared-memory used for metrics
storage.

Two configuration directives, `slab_size` and `max_metric_name_length`,
are added to configure the size of the shared-memory zone dedicated to
metrics and the maximum length of a metric name, respectively.

[1] docs/adr/005-metrics.md
[2] docs/METRICS.md