feat(proxy-wasm) dispatch errors do not interrupt filter chains
Originally this fixes an issue described in #545, in which a failing
dispatch call may cause a pending call to segfault when it is resumed.

To fix this issue, this commit updates how dispatch calls are treated
and makes the execution continue when they fail, including the Lua
bridge resolver (very tricky).